# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Nudge API
  description: >
    Project CRUD routes served by the API Lambda (cmd/api) and the local dev
    server (cmd/local). Every route requires a Cognito access token
    (architecture.md §12/§14, #81) — the caller's user ID comes from the
    token's verified `sub` claim, never from a path, query parameter, or
    request body.
  version: "0.2.0"
servers:
  - url: http://localhost:8080
    description: Local dev server (make local)
security:
  - bearerAuth: []
paths:
...
  /projects/{id}:
    parameters:
      - $ref: "#/components/parameters/ProjectIdPath"
    get:
      summary: Get a project by id
      operationId: getProject
      responses:
        "200":
          description: The project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    put:
      summary: Update a project
      description: >
        Updates the mutable fields only — goal, deadline, constraints,
        status. Name is immutable via this route.
      operationId: updateProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectRequest"
      responses:
        "200":
          description: The updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          description: Invalid body, or failed validation (missing version)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: >
            version doesn't match the project's current version (optimistic
            concurrency, architecture.md §8) — re-fetch and retry, don't
            retry the same request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      summary: Delete a project
      operationId: deleteProject
      responses:
        "204":
          description: Project deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /projects/{id}/tasks:
    parameters:
      - $ref: "#/components/parameters/ProjectIdPath"
    get:
      summary: List a project's tasks
      description: >
        Returns the project's tasks flat, linked by parent_id — not nested.
        Building a tree out of the flat list, if a caller wants one, is
        client-side presentation logic (#125).
      operationId: listTasks
      responses:
        "200":
          description: The project's tasks (empty array if none)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        A Cognito access token. API Gateway's Cognito JWT authorizer
        validates it before the request reaches the API Lambda
        (architecture.md §14) — the user ID is the token's `sub` claim,
        never a path, query, or body value.
  parameters:
    ProjectIdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
  responses:
    NotFound:
      description: No project with this id in this user's partition
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing, invalid, or expired access token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    Project:
      type: object
      properties:
        user_id:
          type: string
        id:
          type: string
        name:
          type: string
        goal:
          type: string
        deadline:
          type: string
          format: date-time
          nullable: true
        constraints:
          type: array
          items:
            type: string
        status:
          type: string
        version:
          type: integer
          description: >
            Optimistic-concurrency version (architecture.md §8). Starts at 1
            on create, increments by 1 on every successful update.
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - user_id
        - id
        - name
        - goal
        - constraints
        - status
        - version
        - created_at
        - updated_at
...
    UpdateProjectRequest:
      type: object
      description: name is intentionally absent — this route never touches it.
      properties:
        version:
          type: integer
          description: >
            Required — the project's version as last read (Project.version).
            Rejected with 409 if it no longer matches the stored version
            (architecture.md §8).
        goal:
          type: string
        deadline:
          type: string
          format: date-time
          nullable: true
        constraints:
          type: array
          items:
            type: string
        status:
          type: string
      required:
        - version
...
    Task:
      type: object
      properties:
        id:
          type: string
        project_id:
          type: string
        parent_id:
          type: string
        title:
          type: string
        description:
          type: string
        status:
          type: string
        order:
          type: integer
        acceptance_criteria:
          type: array
          items:
            type: string
      required:
        - id
        - project_id
        - title
        - status
        - order
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    ValidationErrorResponse:
      type: object
      properties:
        error:
          type: string
        fields:
          type: object
          additionalProperties:
            type: string
      required:
        - error
        - fields